New version of pycheckdoc using ASTs instead of importing modules
Abstract Syntax Trees are used to get documentation info in modules
passed as arguments without having to import the modules.
This version does not skip modules that don't have if __name__ ==
"__main__" guard since there's no execution of the source code being
checked.
A process pool from Pebble is used when reading and parsing the modules
to reduce time taken when multiple modules are being checked.